outsource markdown4k stuff to common4k

+++ common4k/src/main/kotlin/com/github/christophpickl/kpotpourri/common/file/file.kt
@@ -8,9 +8,9 @@ import java.nio.file.Files
 private val log = LOG {}
 
 /**
- * Throws an exception if that file does not exist.
+ * Checks if the File exists or throws a [KPotpourriException].
  *
- * formerly ensureExists()
+ * Formerly known as `ensureExists()`.
  */
 fun File.verifyExists() = this.apply {
     if (!exists()) {
@@ -18,19 +18,31 @@ fun File.verifyExists() = this.apply {
     }
 }
 
+/**
+ * Checks if it's actually a file and not a directory or throws a [KPotpourriException].
+ */
 fun File.verifyIsFile() = this.apply {
     if (!isFile) {
         throw KPotpourriException("Expected to be a file: $absolutePath")
     }
 }
 
+/**
+ * Checks if it's an existing file and not a directory or throws a [KPotpourriException].
+ */
 fun File.verifyExistsAndIsFile() = this.apply {
     verifyExists()
     verifyIsFile()
 }
 
 /**
- * Move a file by using JDK7's Files class.
+ * Cuts off the starting part, e.g.: "/top/sub/file.txt".nameStartingFrom("/top") => "/sub/file.txt"
+ */
+fun File.nameStartingFrom(other: File) =
+        this.canonicalPath.substring(other.canonicalPath.length)
+
+/**
+ * Move a file by using JDK7's [Files] class.
  */
 // MINOR test me
 fun File.move(target: File) {
@@ -38,8 +50,14 @@ fun File.move(target: File) {
     Files.move(this.toPath(), target.toPath())
 }
 
+/**
+ * Delete the directory recursively and recreate the directory structure.
+ */
 fun File.resetDirectory(): File {
     log.debug { "Delete and recreate directory: $canonicalPath" }
+    if (!isDirectory) {
+        throw KPotpourriException("Expected to be a directory: $canonicalPath")
+    }
     if (!deleteRecursively()) {
         throw KPotpourriException("Could not delete directory: $canonicalPath")
     }
@@ -47,6 +65,9 @@ fun File.resetDirectory(): File {
     return this
 }
 
+/**
+ * Creates all directories if not yet existing.
+ */
 fun File.mkdirsIfNecessary(): File {
     if (!exists()) {
         if (mkdirs()) {
@@ -60,4 +81,26 @@ fun File.mkdirsIfNecessary(): File {
         log.debug { "Directory already exists at: $canonicalPath" }
     }
     return this
+}
+
+/**
+ * List all contained files and returns only if suffix matches, ignoring folders by name optionally.
+ */
+fun File.scanForFilesRecursively(
+        suffix: String,
+        ignoreFolders: List<String> = emptyList()
+): List<File> {
+    if (!exists() || !isDirectory) {
+        throw KPotpourriException("Invalid folder: $canonicalPath")
+    }
+    val foundFiles = mutableListOf<File>()
+    listFiles { _ -> true }.forEach { file ->
+        if (file.isDirectory && !ignoreFolders.contains(file.name)) {
+            foundFiles.addAll(file.scanForFilesRecursively(suffix, ignoreFolders))
+        } else if (file.isFile && file.name.endsWith(".$suffix", ignoreCase = true)) {
+            log.trace { "Found proper file while scanning: ${file.canonicalPath}" }
+            foundFiles += file
+        }
+    }
+    return foundFiles
 }
